Fixed W_Help scrolling with one hell of a bodge.
I found that the text area had cached the text of the previous text stored within it, and that was used when calculating the size instead of the new text.

I also found that it has no way of manually invalidating this cache.

A bodge is implemented, read the comments added in this commit to see how it works.

also :
- Correctly breakdown children in the tree, stopping the blank entries being created on every new game.
- Stop user from being able to edit the text area.

- Don't crash if no node is selected in the tree. This happened when clearing the children, as the listener was called on the change, but there was now no children.